Refactor: all: define acceptable glib API verions
glib allows defining the minimum and maximum API versions to use
(i.e. compiler warnings will be generated if an API deprecated before the
minimum version is used, or an API introduced after the maximum version is
used).

Typically these should be defined the same, which we do, to our minimum glib
dependency 2.42.