Build: Always use --no-discard-stderr when running help2man.
We need it in the case where there's not a man page include file for a
tool that's been converted to using glib.  In this case, crm_standby
calls through to crm_attribute when given --version.  The version
information is written to stderr (due to pcmk__cli_help), but there's no
man page include file.  help2man sees no version output where it's
expected and errors.