Feature: CIB: allow resource-stickiness default to be set at build time
A new configure script option, --with-resource-stickiness-default, can now be
used to set a value that will automatically be added to newly generated CIBs
as a resource default for resource-stickiness.

This is done via the Pacemaker build, rather than just let higher-level tools
handle it, so it can be effective regardless of how the cluster is created, and
because a cluster may be created without being started (and thus the CIB is not
available).

An explicit default is used rather than changing the implicit default, to avoid
a plethora of problems with the latter approach (unpredictable behavior in
mixed-version clusters including rolling upgrades, no way for the user to query
implicit defaults currently, no simple way for the user to override the change
due to clone stickiness issues, scheduler regression test issues, etc.).