Feature: controller: support OCF 1.1 reload-agent action
Pacemaker previously used the "reload" action to reload agent parameters, but
most agents used it to reload the service configuration. Pacemaker also misused
the OCF 1.0 "unique" parameter attribute to indicate reloadability.

OCF 1.1 created the "reload-agent" action and "reloadable" parameter attribute
for the Pacemaker usage.

Pacemaker now supports the OCF 1.1 usage. The old usage is now deprecated, but
will be supported if the agent does not claim OCF 1.1 or later compliance and
does not advertise the reload-agent action.