Refactor: tools,libpe_status: Move build_uname_list and build_rsc_list.
These two functions are used to build up lists of objects that we might
want to print out or filter on, based upon the command line.  They are
potentially more useful than in crm_mon where they've been hiding out.
However, they use various pe_ functions to figure it out so they need to
live here and not with the other cmdline code.