Build: extra: remove showscores.sh
The file hasn't been updated more than incidentally since 2010, and it's
currently broken due to changed function names, etc. that it greps for.

Fixing it would ideally involve parsing the XML output from crm_simulate,
but since there doesn't seem to be demand for it, just remove it. If someone
wants to resurrect it and fix it, they're welcome to.