Refactor: tools: Remove uses of mon_op_fence_history.
How much fencing to show can be inferred from the status of the show
flags.  Unfortunately, mon_op_fence_history is not entirely redundant so
I have had to add a new fence_history boolean in crm_mon.c to take its
place.  Basically, mon_op_fence_history was an additional control for
whether or not any fencing history should be shown, while the show_*
flags controlled which parts of that history should be shown.

This can most obviously be seen in running cts/cts-cli, which can be
tested on the command line as:

$ CIB_file=cts/cli/crm_mon.xml tools/crm_mon -1 --include=all

Because this is running with a CIB file, fencing should be disabled (see
the switch on cib->variant).  At the same time, --include=all means the
user is requesting all fencing be shown.  The resulting include/exclude
string is:

--include=fencing-failed,fencing-pending --exclude=fencing --include=all

User requests always go after what was set up automatically.  With the
mon_op_fence_history variable, this include/exclude string didn't
matter.  That boolean was still false, so fencing was not shown.  If we
try to decide whether to display fencing or not based on just the
show_* flags, --include=all is going to override everything and we'll
get error messages in the output about not being able to connect to the
fence agent.

I don't see a better way around this at the moment, thus the new
fence_history boolean.