Refactor: tools,libpe_status: Don't export all the output messages.
These in general do not actually need to be exported.  As long as
there's a default message and pe__register_messages is called, crm_mon's
curses mode will be able to access these.

Note that there is no default handler for maint-mode.  We only want to
display this for text and console formatters.  For other formatters, we
want it to be ignored.