Doc: liblrmd: correct documentation of liblrmd API calls
The liblrmd API documentation was incorrect in a few places.

The connect() and poke() documentation said that an error return would be
reported as -1, when in fact it would usually (and after recent changes,
always) be a negative legacy return code.

unregister_rsc() claimed that the return value would be -1 when operations are
pending on the resource, but actually it is -EINPROGRESS.

In apparent copy-paste errors, cancel() claimed the operation was both
synchronous and asynchronous, and exec_alert() wrongly claimed a successful
result would get a call ID.

get_metadata() and get_metadata_params() referred to a nonexistent lrmd_ok.