Feature: tools: Add an agent-status message for crm_resource.
This moves what was previously only done in an out->info call to its own
output message, which means it will appear in XML output as well.  Also,
note that if --class/--agent/--provider are given, the resource name
will be set to "test".  In that case, do not display the resource name
in the output.

This message will be used for --validate and the --force-* command line
options to crm_resource.

See: rhbz#1644628