Low: tools: Return ENOTUNIQ in find_resource_attr.
We were returning EINVAL in the case where multiple attributes were
found with a given name, but there's a more fitting error code that can
be used.  This also starts to take care of a bunch of other functions
that could be returning EINVAL due to calling find_resource_attr.