Low: controller: set exit reason when remote connection start fails
This allows the user to see something in status besides "Unknown Error".

As of this commit, it's not much more useful, since pcmk__connect_remote()
always passes ENOTCONN for any getaddrinfo() error, meaning the exit reason
will always be "Transport endpoint not connected". However that can be improved
later, and the changes will automatically be picked up in the status.