Build: Don't run "make check" in the doc and extra subdirs.
Because the "check" target normally requires the "all" target to be
built first, this means documentation will be built the first time we
run the tests.  We don't really want to do this - it's weird to have
something happen besides the tests, and it screws up code coverage
numbers.

So, just add a check-recursive target to both directories that does
nothing.  This target will still be run, but because it overrides the
default one, the "all" target will not be built.

There are two downsides to this.  First, we'll get a warning message
from make about overriding a target.  That's probably okay because it'll
remind us of what we're doing.  Second, if we ever introduce tests to
these directories, we'll need to remember to change this patch.