Refactor: controller,liblrmd,libpacemaker: functionize setting event result
lrmd_event_data_t is public API, so we can't refactor it to use
pcmk__action_result_t at this time.

However, we can create equivalent accessor functions for the result members.
This encourages setting agent exit status, execution status, and exit reason
together whenever possible, and will make it easier to refactor to use
pcmk__action_result_t in the future.

This is purely a refactor; no additional exit reasons are set as of this
commit.