Refactor: executor: use pcmk__action_result_t in lrmd_cmd_t
... along with its accessor functions. This is purely a refactor; no additional
exit reasons are set as of this commit.

While the behavior remains the same, there is a slight change in
implementation. Previously, lrmd_cmd_t only stored one set of output when the
command completed: stderr if present, otherwise stdout. Now, the result member
stores both, and the proper one is selected when used. This will increase
memory usage in some situations, but it should be temporary.