Feature: tools: Add an overridden params output message.
This also replaces what was previously being done in an out->info call
with an output message.  This means it shows up in XML output as well.
Also, note that if --class/--agent/--provider are given, the resource
name will be set to "test".  In that case, do not display the resource
name in the output.

See: rhbz#1644628