Test: cts-cli: Add a new test_assert_validate function.
This turns the existing test_assert into _test_assert, and then
test_assert and test_assert_validate are just wrappers around that
function that add the correct extra argument.

For the moment this doesn't do anything.  The point of adding this now
is to ensure that the test output didn't change.