Test: cts: Convert a couple crm_resource tests to XML output.
This is just to get some more coverage of the XML schema.  We're already
covering a bunch of the crm_resource XML output with the loop over the
constraints.xml file.  For the most part, our other uses of crm_resource
just produce the bare minimum of output in XML.