Refactor: libcrmservice: add exit reason to svc_action_t private data
svc_action_t is public API, so we can't refactor it to use
pcmk__action_result_t at this time.

However, we can add an exit reason to its private data member, along with
accessor functions services__exit_reason() and services__set_result(). This
encourages setting action exit status, execution status, and exit reason
together when possible.

The new member and accessors are not used anywhere as of this commit.