Feature: tools: Add an output message for resource actions.
This wraps up the override and agent-status messages into a single
message, along with any stdout/stderr from the resource action.  This
message should be called after taking the action.

This also implements handling XML output from resource actions.  Check
to see if the validate-all action returns XML.  If so, output it as a
CDATA block under a "command" element.  If not, treat it as plain text
and output it as stdout/stderr from a command.

See: rhbz#1644628