Fix: pacemaker-attrd: avoid repeated unfencing of remote nodes
The attribute manager can't record a remote node's attributes to the CIB until
it knows the node is remote. Normally, this is learned when the remote node
starts, because the controller clears the CRM_OP_PROBED attribute and indicates
that it is for a remote node.

However, if a cluster node is down when a remote node starts, and later comes
up, it learns the remote node's existing attributes as part of the attribute
sync. Previously, this did not include whether each value is for a cluster or
remote node, so the newly joined attribute manager couldn't write out remote
nodes' attributes until it learned that via some other event -- which might not
happen before the node becomes DC, in which case its scheduler will not see any
unfencing-related node attributes and may wrongly schedule unfencing.

The sync response handling already calls attrd_lookup_or_create_value(), which
checks PCMK__XA_ATTR_IS_REMOTE, so all we need to do is add that to the sync
response.