Low: pacemaker-attrd: ensure node ID is only set for attributes when known
In most cases, attribute updates contained the node ID, and the node ID was
used by other code, only if known (i.e. positive). However a couple places did
not check this, so add that.

I am unsure whether the missing check caused problems in practice, but there
appears to be the possibility that a remote node would wrongly be added to the
cluster node cache.