Purge remote node attribute when the node is up.
When a remote node is up again after being killed, cleans the attrd's
cache for the node. This is necessary to ensure that the CIB and the
attrd cache will match. Otherwise, when the node is up the cache will be
populated with the values from other nodes (e.g. transient attributes)
and the attributes will be clean from the CIB. Causing a discrepancy
among these data sources.