Fix: scheduler: Display resources that are started but unmanaged.
Certain clone resources are being hidden from display, or are being
shown just as stopped.  This is because those resources are not at least
partially active, so they fall into one of the various tests in
pe__clone_{text,html} but never get print_full set.

So, make sure to set print_full to TRUE for those resources that are not
partially active but are started.