Refactor: CTS: eliminate dead code
The ClusterManager class was originally designed as an abstract class to be
subclassed for each supported cluster stack. A class for the heartbeat stack,
which was also used as a base class for other stacks, was renamed crm_common
in Pacemaker 2.0.0 when heartbeat support was dropped.

crm_common is now the only implementation of ClusterManager, and crm_corosync
is a subclass of crm_common.

It would have made sense to merge crm_common into ClusterManager at the time,
but that was more work than feasible on that deadline. However that left dead
code in the ClusterManager class that was always overridden by crm_common, and
makes reading the code confusing.

This is the first step to merging the classes, removing the ClusterManager
methods that are always overridden.