Low: scheduler: Tweak resource filtering for groups.
Do not display the inactive members of a partially active group by
default.  Require "-r" to be given to crm_mon to make that happen.
However, if the inactive resource is specifically asked for with
"--resource=", display it.  Add test cases for this.

This additionally fixes a bug where inactive children of a partially
active group resource could still be displayed even when not asked for.
This is possible because the filtering list of "*" was taken into
account before considering whether we are displaying inactive resources.