Low: scheduler: Mark partially active groups in output.
This is to help the user see that a group may not be displaying all its
members because those members are inactive.  The count of inactive
resources is only displayed if we are not displaying inactive resources.
If we are, then all resources will be shown so seeing the count is just
extra information.