Low: scheduler: don't make optional constraints with group members mandatory
enum pe_restart (a.k.a. the restart-type resource meta-attribute) is deprecated
and will be removed in a future release, at which time the pe_restart_restart
behavior will disappear, and the pe_restart_ignore behavior (which is the
default) will be the only behavior.

The only difference in pe_restart_restart behavior is for optional ordering
constraints in unpack_simple_rsc_order(), in which case pe_order_implies_then
will be set (and pe_order_implies_first on a symmetric constraint's inverse).
This effectively makes the constraint mandatory.

Previously, groups always set pe_restart_restart on members other than the
first. Only user constraints can possibly be affected by pe_restart_restart, so
it mattered only when a user explicitly created an optional ordering constraint
with an individual group member after the first, making the constraint
mandatory -- which was undocumented and most likely unexpected.