Low: scheduler: Fix list errors in bundle HTML output.
Get rid of the unconditional <li> push and pop, which sometimes caused
the last element of the bundle to be unindented relative to all the
others, and then every list after it to also be unindented.

Then also turn the "Replica[]" line into a list header that all implicit
resources can be chldren of.  This gets rid of some additional list
nesting showing up that looked pretty weird.