Low: scheduler: Fix group-by-node resource HTML output.
For regular output of a list of nodes, everything stays the same.

For the node list with resources grouped under each node, the HTML
output is now different.  Previously, the nested list output was
confusing to read due to extra layers of indentation and blank list
headers.

Now, each node is its own new list with resources underneath as a
sub-list.  The status of the node (online, offline, etc.) is moved from
the node's heading to a separate "Status:" item underneath.  This makes
it easier to construct the resources list.