Refactor: scheduler: tweak pe_order_restart handling
a6af4a04 fixed an issue by limiting pe_order_restart's "unrunnable 'then'"
handling to start actions, to handle the case where a "stop A then start A"
ordering had an unrunnable start. This achieves the same result by checking
whether the ordering is for the same resource instead.

This is logically closer to the intention, i.e. an unrunnable 'then' should
make 'first' unrunnable only if it's for the same resource. There is no way
currently that pe_order_restart could be used for different resources except in
implicit group member orderings, but that might change, and the new test is a
better way to identify the desired case.