Low: libcrmcommon: Also allow string arguments that start with a dash.
There's various places where an option to a command line argument could
itself be a valid command line argument.  For instance:

    crm_attribute -n crm_mon_options -v "-1i3"

The previous patching to pcmk__cmdline_preproc did not take this into
account.  With this patch, options that are last in a string (or by
themselves) and take an argument will have the next command line option
grabbed and copied straight through without processing.

Regression in 2.1.0 caused by a long-standing bug in pcmk__cmdline_preproc.