Refactor: scheduler: drop dead code in pe_action_set_flag_reason()
The function was designed to handle more cases in the future, but in
preparation for an overhaul, simplify it to handle only current callers' cases.

All current callers pass a single flag that is either pe_action_optional,
pe_action_runnable, or pe_action_migrate_runnable, and no caller sets both
reason and text.