Test: tools: Add a test for active resources on offline nodes.
This has to be a crm_mon test, not a scheduler test.  The scheduler
tests always display inactive resources, so we won't be able to know if
the original bug has come back or not.

I've also chosen to use a separate CIB file for testing this.  I'm
trying to keep the number of things tested in a single CIB file to a
minimum, and this seems like a special enough case to where it should be
on its own.