Fix: scheduler: Sort brief output of resources.
If there's a resource type with multiple instances running on multiple
nodes, the list of nodes is not sorted.  This means the output can vary
from host to host.  We have not run into this in previous testing
because it only turns up in brief output, and we apparently do not have
any other test cases in CTS involving more than one instance of a
resource type.