Test: libcrmcommon: Add unit tests for crm_get_msec.
Note that it is unclear from the function documentation what happens if
you do not provide a unit.  The code allows for this possibility, in
which case the unit defaults to seconds.  I have updated the docs for
that.

Also, note that it looks like it will handle numbers with decimal points
in them (given the definition of NUMCHARS), but the type of everything
is long long so there's no way to usefully specify a decimal point.