Refactor: tools: Use the cluster-status message in crm_simulate.
...instead of essentially reimplementing that message.  This reduces
print_cluster_status down to setting flags and calling the message.

A couple scheduler tests have output changes - an empty list of
resources is now shown more often.  This is because by calling the
cluster-status message directly, we do not have the opportunity to pass
FALSE as the print_summary argument to the resource-list message.  This
argument was really only there to support not printing an empty list in
XML.

Also, note that we're still displaying the maintenance mode message even
though this is part of cluster-status.  The reason is we want to keep
its position in scheduler test output where it is.  Moving it breaks pcs
tests, and with the new pcmk_section_maint_mode flag it's easy to keep
everything as it was.