Refactor: libpacemaker: combine 3 files into pcmk_graph_consumer.c
The source files related to transition graphs have gotten somewhat obscure, so
combine the code related to consuming transition graphs into one file.

The contents were changed as minimally as possible: consolidating the copyright
notice and include files, making a function static that now can be, and
dropping one function that became trivial.

A lot of this code could be moved to the controller since very little is needed
elsewhere. I think the original goal was for this to be public API but I don't
see a need for that.