Feature: executor: set exit reason for stonith resource action failures
Previously, a stonith API rc would be mapped to an OCF exit code, which is all
that would be provided to users in status displays. Now, set the string
equivalent of the rc as the exit reason, to give more detail.

This applies to resource actions such as start, stop, and monitor (not actual
fencing actions).

Currently, the only possible values (from svc_action_to_errno()) are ETIME,
ENODATA, and EOPNOTSUPP, but it's better than nothing.