Build: Remove unneeded glib unit test files.
Strip down mk/glib-tap.mk to almost nothing.  All it's really there for
now is to define LOG_DRIVER and LOG_COMPILER.  We could add other test
related stuff to it later if needed.  We didn't need any of the
installed vs. uninstalled unit test stuff, nor the test_data and
dist_test_data stuff.  We only support running unit tests from the
source tree.

Because it doesn't really do anything anymore, rename it to mk/tap.mk.
With that gone, we no longer need support for the ./configure options
that were set up by m4/glibtests.m4, so that entire file can go.  And
then with that gone, we don't need to call GLIB_TESTS anymore.

Then, all the test directory Makefile.am files need to be changed.  Get
rid of the test_data and dist_test_data blocks.  Remove the comment
referring to glib documentation.  Put all the tests into check_PROGRAMS
and then set TESTS equal to that.  glib-tap.mk was doing that for us,
but since we no longer have that file we need to do it manually.