Maint: remove mocked daemon code
This has not been used since its initial creation, and has likely diverged from
the actual daemon code. It is likely that any mocked daemon will be a one-off
based on the code existing at that time. It might be useful to document the
process for creating such a daemon.