Low: scheduler: use correct action when determining order set flags
This is not even worth a regression test since the change is unlikely to affect
any real-world constraints.

The affected code is only used for mandatory, symmetric ordering between two
resource sets. The change would only make a difference when the first resource
set has action set to start or promote, and the second resource set has action
set to stop or demote.