API: libcrmcommon,libcrmservice: deprecate PCMK_OCF_EXEC_ERROR
libcrmservice used this as the agent exit status when the agent could not be
executed. However, the execution status is the proper place to return this
information, not the agent exit status, and in any case, the executor maps
this result to PCMK_OCF_UNKNOWN_ERROR before sending results back to its
clients.