API: libcrmcommon,libcrmservice: deprecate PCMK_OCF_SIGNAL
libcrmservice used this as the agent exit status when the agent was interrupted
by a signal. However, the execution status is the proper place to return this
information, not the agent exit status, and in any case, the executor maps
this result to PCMK_OCF_UNKNOWN_ERROR before sending results back to its
clients.