Refactor: libcrmservice: use PCMK_EXEC_PENDING for multi-stage actions
Certain systemd and nagios actions require a separate monitor action to be
performed as part of them. Previously, we used the custom exit status
PCMK_OCF_PENDING for this purpose. However, this is more properly an execution
status, so replace it with PCMK_EXEC_PENDING instead.

This was never visible to callers, since the exit status would be changed to a
final result before returning it.