Low: executor,controller,libcrmservice: use PCMK_OCF_UNKNOWN_ERROR for timeouts
... instead of PCMK_OCF_TIMEOUT, which is a nonstandard extension.

I originally thought this couldn't be done for backward compatibility,
but the scheduler never checked for the value, and the only thing that did was
libstonithd's svc_action_to_errno(), which has since been updated to correctly
check for PCMK_EXEC_TIMEOUT instead. That was only used in functions called by
the fencer, and for local meta-data/validate calls, so even older remote nodes
should be fine.