Low: executor,libstonithd,libcrmservice: better error handling for generic actions
Previously, services_action_create_generic() did not check for string memory
errors, and logged an error but otherwise ignored a request with too many
arguments, which would lead to executing an external command with partial
arguments.

Now, set the action result to error in such cases, and have the callers check
for it.