Refactor: libcrmservice: functionize getting standard-appropriate exit codes
Previously, services__handle_exec_error() checked the action standard to
determine the proper codes for generic errors, not installed, and insufficient
privileges, then set the code appropriate to the errno.

Now, there are new functions to check the standard for each type of error,
and services__handle_exec_error() uses them.

This will allow the individual functions to be used in other places.