Refactor: fencing: avoid unnecessary use of glib types in stonith_action_t
GPid is equivalent to int on everything but Windows, but it is intended only
for use with g_spawn functions we don't use, so just use int directly instead
(especially since that's what svc_action_t uses, which is where
stonith_action_t gets it from).

Also, use int instead of gint, and void* instead of gpointer, since we're not
using those values with glib functions.